Change value attribute to data value attribute for calculator buttons
If the calculator buttons have a value attribute then change it to data
value attribute since these buttons are actually <div> elements and the
value attribute can only be used for <button>,<input>,<meter>,<li>,
<option>,<progress> and <param> elements.

The data-* attribute adds custom information to a <div> element.
The * part is replaced with a lowercase string, such as data-id,
data-source, data-category,data-value etc.